+++ src/components/PasswordStrength/PasswordStrength.styles.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+
+export const ProgressBarContainer = styled.div`
+	width: 80px;
+`